requirements.txt
python_version >= '3.8'
Flask==3.1.2
flask-bcrypt==1.0.1
pymysql==1.1.1
python-dotenv==1.1.1
openai==1.91.1
